CLI: rename modules
