+++ gr-utils/python/modtool/cli/base.py
@@ -48,8 +48,8 @@ class CommandCLI(click.Group):
         """
         cmds = []
         for filename in os.listdir(self.cmd_folder):
-            if filename.endswith('.py') and filename.startswith('cli_'):
-                cmds.append(filename[4:-3])
+            if filename.endswith('.py') and not '_' in filename:
+                cmds.append(filename[:-3])
         cmds.remove('base')
         cmds += self.commands
         return sorted(cmds)
@@ -62,7 +62,7 @@ class CommandCLI(click.Group):
         try:
             if sys.version_info[0] == 2:
                 cmd_name = cmd_name.encode('ascii', 'replace')
-            mod = import_module('gnuradio.modtool.cli.cli_' + cmd_name)
+            mod = import_module('gnuradio.modtool.cli.' + cmd_name)
         except ImportError:
             return self.commands.get(cmd_name)
         return mod.cli
